# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"pf-be-java-hisp-w20-g07","description":"pf-be-java-hisp-w20-g07","contact":{"name":"manuel diaz","email":"manuel.diaz@mercadolibre.com.co"},"version":"last"},"servers":[{"url":"http://localhost:8080","description":"Scope local"}],"paths":{"/api/v1/fresh-products/orders/{orderId}":{"get":{"tags":["product-controller"],"operationId":"findProductsByOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductOrderResponseDTO"}}}}}}},"put":{"tags":["product-controller"],"operationId":"updateOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/inboundorder":{"put":{"tags":["product-controller"],"operationId":"r1__2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDto"}}}}}},"post":{"tags":["product-controller"],"operationId":"r1__1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDto"}}}}}}},"/api/v1/log-in":{"post":{"tags":["session-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/api/v1/fresh-products/orders":{"post":{"tags":["product-controller"],"operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/{idProduct}/warehouse/list":{"get":{"tags":["product-controller"],"operationId":"r4__1","parameters":[{"name":"idProduct","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockResponseDto"}}}}}}},"/api/v1/fresh-products/{idProduct}/batch/list":{"get":{"tags":["product-controller"],"operationId":"getProductInStock","parameters":[{"name":"idProduct","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"order","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockDTO"}}}}}}},"/api/v1/fresh-products/list":{"get":{"tags":["product-controller"],"operationId":"findProductByCategory","parameters":[{"name":"code","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponseDTO"}}}}}}}},"/api/v1/fresh-products/batch/list/due-date/{cantDays}":{"get":{"tags":["product-controller"],"operationId":"r5__1_1","parameters":[{"name":"cantDays","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}},{"name":"order","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FindBatchesDueToExpireSoonDto"}}}}}}}},"components":{"schemas":{"OrderStatusDTO":{"required":["status_code"],"type":"object","properties":{"status_code":{"type":"string"}}},"ProductDTO":{"required":["product_id","quantity"],"type":"object","properties":{"product_id":{"minimum":1,"type":"integer","format":"int32"},"quantity":{"minimum":0,"type":"integer","format":"int32"}}},"PurchaseOrderRequestDTO":{"required":["buyer_id","date","order_status","product"],"type":"object","properties":{"date":{"type":"string","format":"date"},"buyer_id":{"type":"integer","format":"int32"},"order_status":{"$ref":"#/components/schemas/OrderStatusDTO"},"product":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}},"BatchDto":{"required":["batch_number","current_quantity","current_temperature","due_date","initial_quantity","manufacturing_date","manufacturing_time","minimum_temperature","product_id"],"type":"object","properties":{"batch_number":{"minimum":1,"type":"integer","format":"int32"},"product_id":{"minimum":1,"type":"integer","format":"int32"},"current_temperature":{"type":"number","format":"double"},"minimum_temperature":{"type":"number","format":"double"},"initial_quantity":{"minimum":1,"type":"integer","format":"int32"},"current_quantity":{"minimum":0,"type":"integer","format":"int32"},"manufacturing_date":{"type":"string","format":"date"},"manufacturing_time":{"type":"string","format":"date-time"},"due_date":{"type":"string","format":"date"}}},"InboundOrderDto":{"required":["batch_stock","date","id","section"],"type":"object","properties":{"id":{"minimum":1,"type":"integer","format":"int32"},"date":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/SectionDto"},"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDto"}}}},"InboundOrderRequestDto":{"required":["inbound_order"],"type":"object","properties":{"inbound_order":{"$ref":"#/components/schemas/InboundOrderDto"}}},"SectionDto":{"required":["section_code","warehouse_code"],"type":"object","properties":{"section_code":{"minimum":1,"type":"integer","format":"int32"},"warehouse_code":{"minimum":1,"type":"integer","format":"int32"}}},"InboundOrderResponseDto":{"type":"object","properties":{"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDto"}}}},"UserRequestDTO":{"type":"object","properties":{"user_name":{"type":"string"},"password":{"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"user_name":{"type":"string"},"token":{"type":"string"}}},"PurchaseOrderResponseDTO":{"type":"object","properties":{"total_price":{"type":"number","format":"double"}}},"StockResponseDto":{"type":"object","properties":{"product_id":{"type":"integer","format":"int32"},"warehouses":{"type":"array","items":{"$ref":"#/components/schemas/WarehouseStockDto"}}}},"WarehouseStockDto":{"type":"object","properties":{"warehouse_code":{"type":"integer","format":"int32"},"total_quantity":{"type":"integer","format":"int32"}}},"BatchProductDTO":{"type":"object","properties":{"batch_number":{"type":"integer","format":"int32"},"current_quantity":{"type":"integer","format":"int32"},"due_date":{"type":"string","format":"date"}}},"BatchStockDTO":{"type":"object","properties":{"section":{"$ref":"#/components/schemas/SectionDto"},"product_id":{"type":"integer","format":"int32"},"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchProductDTO"}}}},"ProductOrderResponseDTO":{"type":"object","properties":{"product_id":{"type":"integer","format":"int32"},"name":{"type":"string"},"price":{"type":"number","format":"double"},"quantity":{"type":"integer","format":"int32"}}},"ProductResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"price":{"type":"number","format":"double"}}},"FindBatchesDueToExpireSoonDto":{"type":"object","properties":{"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDto"}}}}}}}